ui: refactor; move renamedStatuses to separate file
Release note: none
