+++ pkg/ui/src/views/jobs/jobStatusOptions.ts
@@ -1,4 +1,6 @@
 import {cockroach} from "src/js/protos";
+import _ from "lodash";
+import {JOB_STATUS_CANCELED, JOB_STATUS_FAILED, JOB_STATUS_SUCCEEDED} from "src/views/jobs/index";
 import Job = cockroach.server.serverpb.JobsResponse.IJob;
 
 export const statusOptions = [
@@ -14,3 +16,16 @@ export const statusOptions = [
 export function jobHasOneOfStatuses(job: Job, ...statuses: string[]) {
   return statuses.indexOf(job.status) !== -1;
 }
+
+export const renamedStatuses = (status: string) => {
+  switch (status) {
+    case JOB_STATUS_SUCCEEDED:
+      return "Completed";
+    case JOB_STATUS_FAILED:
+      return "Import failed";
+    case JOB_STATUS_CANCELED:
+      return "Import failed";
+    default:
+      return _.capitalize(status);
+  }
+};
